+++ kaithem/__init__.py
@@ -4,21 +4,18 @@ from .src import main
 
 # Config keys exactly the same as config.yaml.  Note that not all take effect instantly.
 config = main.config
-api = None
+started = False
 
 
 def initialize_app(cfg: Optional[Dict[str, Any]] = None):
-    """Initialize the app and return an API object
-    identical to the kaithem namespace in web-created modules"""
-    global api
-    if not api:
+    """Initialize the app"""
+    global started
+    if not started:
         main.initialize(cfg)
     else:
         raise RuntimeError("The app has already been set up")
-    from .src.kaithemobj import kaithem
 
-    api = api
-    return kaithem
+    started = True
 
 
 # Call this to start a