+++ kaithem/src/docs/changes.md
@@ -11,6 +11,7 @@ This release moves some of the last few settings to module resources, so that al
 - :coffin: CSS theme setting is gone, use core/css_theme
 - :coffin:  Homepage redirect setting is now core/homepage_redirect
 - :coffin:  No device ever used this, but the on-page text file editor is removed.
+- :coffin: The global kaithem API object is fully gone
 
 #### Others
 - :technologist: Lots of dead code gone, guided by test conerage metrics.